[sql] update char_effects.sql "subid" field to int(10) to match CStatusEffectContainer

+++ sql/char_effects.sql
@@ -18,7 +18,7 @@ CREATE TABLE IF NOT EXISTS `char_effects` (
   `power` smallint(5) unsigned NOT NULL DEFAULT '0',
   `tick` int(10) unsigned NOT NULL DEFAULT '0',
   `duration` int(10) unsigned NOT NULL DEFAULT '0',
-  `subid` smallint(5) unsigned NOT NULL DEFAULT '0',
+  `subid` int(10) unsigned NOT NULL DEFAULT '0',
   `subpower` smallint(5) NOT NULL DEFAULT '0',
   `tier` smallint(5) unsigned NOT NULL DEFAULT '0',
   `flags` int(8) unsigned NOT NULL DEFAULT '0',
